Add administrator privileges

+++ BoxingParadiseBackend/Services/BoxerService.cs
@@ -12,10 +12,12 @@ namespace BoxingParadiseBackend.Services
     public class BoxerService : IBoxerService
     {
         private readonly IBoxerRepository m_BoxerRepository;
+        private readonly IAdministratorService m_AdministratorService;
 
-        public BoxerService(IBoxerRepository boxerRepository)
+        public BoxerService(IBoxerRepository boxerRepository, IAdministratorService administratorService)
         {
             m_BoxerRepository = boxerRepository;
+            m_AdministratorService = administratorService;
         }
 
         public async Task<IList<BoxerDto>> GetBoxers()
@@ -25,14 +27,20 @@ namespace BoxingParadiseBackend.Services
                     .ToList();
         }
 
-        public Task CreateBoxer(BoxerDto boxer)
+        public async Task CreateBoxer(BoxerDto boxer, string adminKey)
         {
-            return m_BoxerRepository.Persist(Mapper.Map<Boxer>(boxer));
+            if (await m_AdministratorService.IsProvidedAdministratorKeyValid(adminKey))
+            {
+                await m_BoxerRepository.Persist(Mapper.Map<Boxer>(boxer));
+            }
         }
 
-        public Task DeleteBoxer(int id)
+        public async Task DeleteBoxer(int id, string adminKey)
         {
-            return m_BoxerRepository.Delete(id);
+            if (await m_AdministratorService.IsProvidedAdministratorKeyValid(adminKey))
+            {
+                await m_BoxerRepository.Delete(id);
+            }
         }
     }
 }
